Unstable switching to OpenCV 2.1
Attempt to switch over to modern OpenCV 2.1 functions, since most of our
stuff was using legacy Python runctions.   This seems to have changed a
few datatypes, so parts of the code still need to be converted and
patched up.   Image capture is much faster but I seem to have thrown off
the stability of our turret control loop.   Not sure why that would be
unless there is some mismatch between our camera movement ending and a
new picture being taken.