Sweep, Guard and profile mode
added new modes for the turret behavior to sweep the range or hold a
position.   Added option to enable detection of side profiles of faces,
which is much slower than full frontal faces only.
tried to standardize across different launchers, but untested on the
launcher1130 so it may not be centering where we want.